+++ game/g_main.c
@@ -337,6 +337,7 @@ vmCvar_t	lmd_bounty_streaks_by;
 vmCvar_t	lmd_skillpoint_cost;
 vmCvar_t	lmd_exp_for_level;
 vmCvar_t    lmd_skillpoints_perlevel;
+vmCvar_t lmd_mastery_level_skills_level;
 								 
 //RoboPhred:
 vmCvar_t lmd_DataPath;
@@ -865,6 +866,8 @@ static cvarTable_t		gameCvarTable[] = {
 { &lmd_skillpoint_cost, "lmd_skillpoint_cost", "150", CVAR_ARCHIVE, 0, qfalse },
 { &lmd_exp_for_level, "lmd_exp_for_level", "450", CVAR_ARCHIVE, 0, qfalse },
 { &lmd_skillpoints_perlevel, "lmd_skillpoints_perlevel", "4", CVAR_ARCHIVE, 0, qfalse },
+{ &lmd_mastery_level_skills_level, "lmd_mastery_level_skills_level", "3", CVAR_ARCHIVE, 0, qfalse },
+
 
 // Lugormod cvars:
 { &g_noVoteTime, "g_noVoteTime", "5", CVAR_ARCHIVE,0, qfalse, qfalse,